requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.0
black==21.11b1
certifi==2020.6.20
chardet==3.0.4
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.7
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-filter==2.4.0
django-import-export==2.1.0
django-seed==0.2.2
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
Faker==4.14.2
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pyparsing==3.0.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
regex==2021.11.10
reportlab==3.6.2
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
text-unidecode==1.3
tomli==1.2.2
typing_extensions==4.0.0
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
